refactor(search): extract shared timing/overfetch helpers into results.py
BACKEND_LEG_TIMING_KEYS and sane_overfetch_factor were private to
daemon.py but three files that survive P12 imported them:
router search.py, federated_search.py, graph_search_service.py.

Moving them into results.py (which is already the shared type home)
so daemon.py can be deleted without breaking the survivors.  Renamed
from private (_BACKEND_LEG_TIMING_KEYS / _sane_overfetch_factor) to
public since they're now cross-module.
